The common tests classes was moved to a separate file.
