+++ tests/test_plugin_titlelink.py
@@ -2,7 +2,7 @@
 import os, sys
 sys.path.append(os.path.abspath(os.path.curdir))
 from plugin_titlelink import *
-
+from common import *
 messages = [
     '',
     'test',
@@ -22,26 +22,7 @@ messages = [
     'http://putin.nash.president.ru',
     ' https://raw.githubusercontent.com/ridouchire/RitsuBot/master/plugin_titlelink.py '
 ]
-room = 'ume@conference.ume.mooo.com'
-nick = 'test'
 
-class Message:
-    def __init__(self, text):
-        self.text = text
-        
-    def getType(self):
-        return 'groupchat'
-
-    def getBody(self):
-        return text
-
-class Bot:
-    self_nick = {}
-    def __init__(self, room, nick):
-        self.self_nick[room] = nick
-    def send_room_message(self, target, res):
-        print('target: {}, message: {}\n'.format(target, res))
-    
 bot = Bot(room, nick)
 
 for text in messages:
